--- packages/types/src/searchFormType.ts
@@ -1,4 +1,0 @@
-export interface KeywordsFormData {
-  keyword: string;
-  searchHistory: string[];
-}
